+++ R/global.R
@@ -82,6 +82,15 @@ utils::globalVariables(
     "pi_upper",
     "df",
     ".y",
-    ".grp"
+    ".grp",
+    "nobs",
+    "logLik",
+    "model",
+    "name",
+    "o_max",
+    "o_min",
+    "R2",
+    "AIC",
+    "p"
   )
 )