+++ NEWS.md
@@ -1,5 +1,16 @@
 # flexFitR (development version)
 
+## New features
+
+* `performance()` function added to evaluate the performance of different models.
+* `plot.performance()` S3 method to plot an object of class `performance`.
+
+## Changes
+
+* `metrics()` returns R2 instead of r_squared.
+
+## Bug fixes
+
 * Fixed conflict of `modeler()` with upcoming version of `future`.
 * Fixed increase dependency to R (>=4.1).
 * Fixed regression function not found in the environment when running in parallel.